requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
pandas>=1.3.0
numpy>=1.21.0
bcrypt>=4.0.0
PyJWT>=2.8.0
redis>=5.0.0
